Fix #708: When prob(up) < 0.5, trader profit is always zero (PR #712)
Fixes #708. The key fix is when the trader exits the trading position in the case of `not pred_up`. 

Related change made: removed Eth stuff from sim_engine. Why: this is _sim_ and doesn't deal with blockchain; the Eth stuff was hurting understandability, including my work to solve this ticket